create inode.depth

+++ src/index.ts
@@ -1,23 +1,36 @@
 import { Graph, digraph } from "graphviz";
 
 class INode {
-  parent: Folder;
+  depth: number = 0;
+  parent: Folder | null = null;
   constructor(public name: string) {}
 
   move(newParent: Folder) {
     console.log(`Moving ${this.name} to ${newParent.name}`);
-    this.parent.removeFile(this);
+    if (this.parent) {
+      this.parent.removeFile(this);
+    }
     newParent.addChild(this);
     this.parent = newParent;
+  }
+
+  updateDepth() {
+    this.depth = this.parent ? this.parent.depth + 1 : 0;
   }
 }
 
 class Folder extends INode {
   children: INode[] = [];
 
+  updateDepth() {
+    super.updateDepth();
+    this.children.forEach((child) => child.updateDepth());
+  }
+
   addChild(inode: INode) {
     this.children.push(inode);
     inode.parent = this;
+    this.updateDepth();
   }
 
   removeFile(inode: INode) {
@@ -51,7 +64,7 @@ class FileNode extends INode {
   resume() {
     return {
       name: this.name,
-      parent: this.parent.name,
+      parent: this.parent?.name,
       imports: this.imports.map((file) => file.name),
       importedBy: this.importedBy.map((file) => file.name),
     };
@@ -78,14 +91,16 @@ file2.addImport(file4);
 
 folder1.addChild(file1);
 folder2.addChild(file2);
-root.addChild(file3);
+folder1.addChild(file3);
 root.addChild(file4);
 
 function optimizeFolderStructure() {
   files.forEach((file) => {
     if (file.imports.length === 0 && file.importedBy.length === 1) {
       const importer = file.importedBy[0];
-      file.move(importer.parent);
+      if (importer.parent) {
+        file.move(importer.parent);
+      }
     }
   });
 }
@@ -97,8 +112,12 @@ console.log(file4.resume());
 function plotFolders(g: Graph, folder: Folder) {
   const cluster = g.addCluster("cluster_" + folder.name);
   cluster.set("label", folder.name);
+  cluster.set("label", folder.name + " " + folder.depth);
+
   folder.files.forEach((file) => {
-    cluster.addNode(file.name);
+    const node = cluster.addNode(file.name);
+    node.set("shape", "box");
+    node.set("label", file.name + " " + file.depth);
   });
   folder.subfolders.forEach((subfolder) => {
     plotFolders(cluster, subfolder);
